tracks/cdaa_qcre/requirements-sqgm.txt
importlib_metadata==8.6.1
networkx==3.4.2
numpy==2.2.3
qiskit-terra==0.46.3
